requirements.txt
Django==4.2.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
djangorestframework==3.14.0
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0
pandas==2.0.3
numpy==1.24.4